netkvm: minimize impact of teaming support on performance
In case the teaming support is not relevant (which is
the case on most of setups), we'd like to minimize to
performance loss due to the fact that we need to check
presense of the teaming on each packet transmit or
packet return. For that we prepare two possibilities:
1) use indication that there is at least one adapter bound
   to the protocol.
2) use acked 'standby' feature
Currently first option is used.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>